chore: move module-side-effects.ts to types
