--- packages/rolldown/src/treeshake/index.ts
@@ -1,1 +1,0 @@
-export * from './module-side-effects'
