+++ packages/rolldown/src/types/module-side-effects.ts
